feat: connection to front end
