+++ web/app/routes/tx.tsx
@@ -1,12 +1,6 @@
 import { ActionFunctionArgs, json, type MetaFunction } from "@remix-run/node";
 import { Form, useActionData } from "@remix-run/react";
-import { useState } from "react";
-import {
-  Button,
-  logCuriosity,
-  RootSection,
-  ValidationAccordion,
-} from "../components";
+import { Button, logCuriosity, RootSection } from "../components";
 import * as server from "./tx.server";
 import TOPICS from "./tx.topics";
 
@@ -32,8 +26,12 @@ export async function action({ request }: ActionFunctionArgs) {
   const raw = formData.get("raw");
 
   if (raw) {
-    const res = server.safeParseTx(raw.toString());
-    return json({ ...res, raw });
+    const { section, validations } = server.safeParseTx(raw.toString());
+    return json({
+      ...section,
+      raw,
+      ...validations,
+    });
   } else {
     return json({ error: "an empty value? seriously?" });
   }
@@ -59,54 +57,9 @@ function ExampleCard(props: { title: string; address: string }) {
 export default function Index() {
   const data = useActionData();
 
-  const [open, setOpen] = useState(false);
-  const handleClick = () => setOpen(!open);
-
   logCuriosity(data);
 
-  const validations: IValidation[] = [
-    { name: "Non empty inputs", value: true, description: "Sucessful" },
-    {
-      name: "All inputs in utxos",
-      value: false,
-      description:
-        "Lorem ipsum dolor sit amet consectetur adipisicing elit. Porro id maiores exercitationem asperiores molestias assumenda doloremque magnam fugit. Iure dolorum fugit facilis autem incidunt vero necessitatibus consectetur ducimus recusandae blanditiis!",
-    },
-    { name: "Validity interval", value: true, description: "Sucessful" },
-    { name: "Fee", value: true, description: "Sucessful" },
-    {
-      name: "Preservation of value",
-      value: false,
-      description:
-        "Lorem ipsum dolor sit amet consectetur adipisicing elit. Porro id maiores exercitationem asperiores molestias assumenda doloremque magnam fugit. Iure dolorum fugit facilis autem incidunt vero necessitatibus consectetur ducimus recusandae blanditiis!",
-    },
-    {
-      name: "Min lovelace per UTxO",
-      value: false,
-      description:
-        "Lorem ipsum dolor sit amet consectetur adipisicing elit. Porro id maiores exercitationem asperiores molestias assumenda doloremque magnam fugit. Iure dolorum fugit facilis autem incidunt vero necessitatibus consectetur ducimus recusandae blanditiis!",
-    },
-    { name: "Output value size", value: true, description: "Successful" },
-    { name: "Network Id", value: true, description: "Successful" },
-    { name: "Tx size", value: true, description: "Successful" },
-    { name: "Tx execution units", value: true, description: "Successful" },
-    { name: "Minting", value: true, description: "Successful" },
-    {
-      name: "Well formed",
-      value: false,
-      description:
-        "Lorem ipsum dolor sit amet consectetur adipisicing elit. Porro id maiores exercitationem asperiores molestias assumenda doloremque magnam fugit. Iure dolorum fugit facilis autem incidunt vero necessitatibus consectetur ducimus recusandae blanditiis!",
-    },
-    { name: "Script witness", value: true, description: "Successful" },
-    { name: "Languages", value: true, description: "Successful" },
-    {
-      name: "Auxiliary data hash",
-      value: false,
-      description:
-        "Lorem ipsum dolor sit amet consectetur adipisicing elit. Porro id maiores exercitationem asperiores molestias assumenda doloremque magnam fugit. Iure dolorum fugit facilis autem incidunt vero necessitatibus consectetur ducimus recusandae blanditiis!",
-    },
-    { name: "Script data hash", value: true, description: "Successful" },
-  ];
+  const validations: IValidation[] = data?.validations || [];
 
   return (
     <main className="mt-10 px-4">
@@ -146,23 +99,7 @@ export default function Index() {
       )}
 
       {!!data && (
-        <div className="flex flex-col">
-          <div className="mb-14">
-            <button
-              className={`flex items-center w-full text-left select-none duration-300`}
-              onClick={handleClick}
-            >
-              <div
-                className={`h-8 w-8 inline-flex items-center justify-center duration-300 `}
-              >
-                {open ? "▼" : "▶"}
-              </div>
-              <h4 className="text-3xl ">Tx Validations</h4>
-            </button>
-            {open && <ValidationAccordion validations={validations} />}
-          </div>
-          <RootSection data={data} topics={TOPICS} />
-        </div>
+        <RootSection data={data} topics={TOPICS} validations={validations} />
       )}
     </main>
   );
